Moved the MemoryManager from the io module to the unsafe module and make it public
Because the page cache might not be the only thing that is interested in allocating alligned memory.
